+++ CHANGELOG.md
@@ -6,11 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Integration into Nextcloud's Files
 - Added license (GNU Affero General Public License)
 
 ### Changed
 
-- This release modernizes the application code in order to be able to utilize more of NextCloud's latest features.
+- Modernized various aspects of the app
 
 ## [V2.0.0] - 2026-06-16